backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
Pillow==10.4.0
transformers==4.44.2
torch==2.4.1
torchvision==0.19.1
faiss-cpu==1.8.0.post1
google-generativeai==0.8.3
requests==2.32.3
python-dotenv==1.0.1
numpy==1.26.4
aiofiles==24.1.0
